fix: Fix 60 FPS crash on Samsung by checking `CamcorderProfile.maxFps`
