+++ package/android/src/main/java/com/mrousavy/camera/core/CameraDeviceDetails.kt
@@ -6,7 +6,9 @@ import android.hardware.camera2.CameraCharacteristics
 import android.hardware.camera2.CameraExtensionCharacteristics
 import android.hardware.camera2.CameraManager
 import android.hardware.camera2.CameraMetadata
+import android.media.CamcorderProfile
 import android.os.Build
+import android.util.Log
 import android.util.Range
 import android.util.Size
 import com.facebook.react.bridge.Arguments
@@ -22,7 +24,9 @@ import com.mrousavy.camera.types.LensFacing
 import com.mrousavy.camera.types.Orientation
 import com.mrousavy.camera.types.PixelFormat
 import com.mrousavy.camera.types.VideoStabilizationMode
+import com.mrousavy.camera.utils.CamcorderProfileUtils
 import kotlin.math.atan2
+import kotlin.math.min
 import kotlin.math.sqrt
 
 @SuppressLint("InlinedApi")
@@ -200,7 +204,11 @@ class CameraDeviceDetails(private val cameraManager: CameraManager, val cameraId
 
     videoSizes.forEach { videoSize ->
       val frameDuration = cameraConfig.getOutputMinFrameDuration(videoFormat, videoSize)
-      val maxFps = (1.0 / (frameDuration.toDouble() / 1_000_000_000)).toInt()
+      var maxFps = (1.0 / (frameDuration.toDouble() / 1_000_000_000)).toInt()
+      val maxEncoderFps = CamcorderProfileUtils.getMaximumFps(cameraId, videoSize)
+      if (maxEncoderFps != null && maxEncoderFps < maxFps) {
+        maxFps = maxEncoderFps
+      }
 
       photoSizes.forEach { photoSize ->
         val map = buildFormatMap(photoSize, videoSize, Range(1, maxFps))
